Сделана функция фомирования шума для генератора AWGN. Протестировано.

+++ source/recievers/POWER_METER.f90
@@ -41,7 +41,7 @@ SUBROUTINE  POWER_METER_REAL(signal,fd,power,period)
    FUNCTION GetSignalRmsPowerINT2(signal,length)
         INTEGER(2),  INTENT(IN)    :: signal(:)
         INTEGER(8),  INTENT(IN)    :: length
-        REAL(4)   ,  ALLOCATABLE   :: tmpArray(:)
+        REAL(8)   ,  ALLOCATABLE   :: tmpArray(:)
         REAL(8)                    :: GetSignalRmsPowerINT2,summ
         ALLOCATE(tmpArray(1:length))
         tmpArray = float(signal(1:length))/32767.0
